+++ src/main/java/com/example/controller/PostController.java
@@ -12,6 +12,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 
+import java.math.BigInteger;
 import java.util.Date;
 import java.util.List;
 
@@ -39,7 +40,7 @@ public class PostController {
 
         if(jwtTokenProvider.validateToken(jwtToken)) {
             Claims claimsFromJwt = jwtTokenProvider.getClaimsFromJwt(jwtToken);
-            Long userID = (Long) claimsFromJwt.get("userID");
+            BigInteger userID = (BigInteger) claimsFromJwt.get("userID");
             String email = (String) claimsFromJwt.get("email");
 
             Connection foundedUser = connectionRepository.findByUserIDAndEmailAndAuthenticated(userID, email, true);
@@ -70,7 +71,7 @@ public class PostController {
 
         if(jwtTokenProvider.validateToken(jwtToken)) {
             Claims claimsFromJwt = jwtTokenProvider.getClaimsFromJwt(jwtToken);
-            Long userID = (Long) claimsFromJwt.get("userID");
+            BigInteger userID = (BigInteger) claimsFromJwt.get("userID");
             String email = (String) claimsFromJwt.get("email");
 
             Connection foundedUser = connectionRepository.findByUserIDAndEmailAndAuthenticated(userID, email, true);
@@ -96,7 +97,7 @@ public class PostController {
     public ResponseEntity<?> updatePostById(@RequestBody PostDto postDto, @PathVariable Long postId, @PathVariable String jwtToken) {
         if(jwtTokenProvider.validateToken(jwtToken)) {
             Claims claimsFromJwt = jwtTokenProvider.getClaimsFromJwt(jwtToken);
-            Long userID = (Long) claimsFromJwt.get("userID");
+            BigInteger userID = (BigInteger) claimsFromJwt.get("userID");
             String email = (String) claimsFromJwt.get("email");
 
             Connection foundedUser = connectionRepository.findByUserIDAndEmailAndAuthenticated(userID, email, true);
